requirements.txt
Flask
PyMuPDF
requests
python-dotenv
google-generativeai
gunicorn